+++ spec/lib/karafka/pro/routing/features/swarm/contracts/routing_spec.rb
@@ -27,6 +27,32 @@ RSpec.describe_current do
     it { expect { validation }.to raise_error(validation_error) }
   end
 
+  context 'when routes in hash do not match all the topics' do
+    before do
+      builder.draw do
+        topic 'topic1' do
+          consumer Class.new(Karafka::BaseConsumer)
+          swarm(nodes: { 1 => [0, 2] })
+        end
+      end
+    end
+
+    it { expect { validation }.to raise_error(validation_error) }
+  end
+
+  context 'when routes hash matches all the topics' do
+    before do
+      builder.draw do
+        topic 'topic1' do
+          consumer Class.new(Karafka::BaseConsumer)
+          swarm(nodes: { 0 => [0], 1 => [1], 2 => [2] })
+        end
+      end
+    end
+
+    it { expect { validation }.not_to raise_error }
+  end
+
   context 'when routes match all the topics' do
     before do
       builder.draw do